+++ app/controllers/actions/IdentifierAction.scala
@@ -17,234 +17,21 @@
 package controllers.actions
 
 import com.google.inject.Inject
-import com.kenshoo.play.metrics.Metrics
-import config.FrontendAppConfig
-import connectors.SubscriptionConnector
-import controllers.actions.IdentifierAction.{pptEnrolmentIdentifierName, pptEnrolmentKey}
-import controllers.{routes => agentRoutes}
-import controllers.home.{routes => homeRoutes}
-import models.Mode.NormalMode
-import models.requests.{IdentifiedRequest, IdentityData}
-import models.SignedInUser
-import models.subscription.LegalEntityDetails
-import play.api.Logger
-import play.api.libs.json.{Json, OFormat}
-import play.api.mvc.Results._
+import models.requests.IdentifiedRequest
 import play.api.mvc._
-import repositories.SessionRepository
-import repositories.SessionRepository.Paths.SubscriptionIsActive
-import uk.gov.hmrc.auth.core._
-import uk.gov.hmrc.auth.core.retrieve.~
-import uk.gov.hmrc.http.HeaderCarrier
-import uk.gov.hmrc.play.http.HeaderCarrierConverter
 
 import scala.concurrent.{ExecutionContext, Future}
-
-object IdentifierAction {
-  val pptEnrolmentKey            = "HMRC-PPT-ORG"
-  val pptEnrolmentIdentifierName = "EtmpRegistrationNumber"
-}
 
 trait IdentifierAction
   extends ActionBuilder[IdentifiedRequest, AnyContent]
 
-
-//mixing action, used by almost everywhere
-class PanAndLindsay @Inject()(
-                               val parser: BodyParsers.Default,
-                               identifierAction: IdentifierActionOld,
-                               subscriptionFilter: SubscriptionFilter
+class IdentifyAndCheckSubscription @Inject()(
+                                              val parser: BodyParsers.Default,
+                                              identifierAction: AuthAction,
+                                              subscriptionFilter: SubscriptionFilter
                              )(implicit val executionContext: ExecutionContext) extends IdentifierAction {
+
   override def invokeBlock[A](request: Request[A], block: IdentifiedRequest[A] => Future[Result]): Future[Result] =
     identifierAction.andThen(subscriptionFilter).invokeBlock(request, block)
-}
-
-
-case class PPTSubscriptionDetails(legalEntityDetails: LegalEntityDetails)
-object PPTSubscriptionDetails {
-  implicit val format: OFormat[PPTSubscriptionDetails] = Json.format[PPTSubscriptionDetails]
-}
-
-class SubscriptionFilter @Inject()(
-                                    subscriptionConnector: SubscriptionConnector,
-                                    sessionRepository: SessionRepository
-                                  )(implicit val executionContext: ExecutionContext)
-  extends ActionFilter[IdentifiedRequest] with HeaderCarrierConverter {
-
-  override protected def filter[A](request: IdentifiedRequest[A]): Future[Option[Result]] = {
-    sessionRepository.get[PPTSubscriptionDetails](request.cacheKey, SubscriptionIsActive).flatMap{
-      case Some(_) => Future.successful(None)
-      case _ =>
-        subscriptionConnector.get(request.pptReference)(fromRequestAndSession(request, request.session)).flatMap{
-          case Right(subscription) => sessionRepository
-            .set(request.cacheKey, SubscriptionIsActive, PPTSubscriptionDetails(subscription.legalEntityDetails))
-            .map(_ => None)
-          case Left(eisFailure) if eisFailure.isDeregistered =>
-            Future.successful(Some(Redirect(agentRoutes.DeregisteredController.onPageLoad()))) //agent?????
-          case Left(eisFailure) =>
-            throw new RuntimeException(
-              s"Failed to get subscription - ${eisFailure.failures.map(_.headOption.map(_.reason))
-                .getOrElse("no underlying reason supplied")}"
-            )
-        }
-    }
-  }
-
-}
-
-
-//this is just auth?
-trait IdentifierActionOld
-    extends ActionBuilder[IdentifiedRequest, AnyContent]
-    with ActionFunction[Request, IdentifiedRequest]
-
-class AuthenticatedIdentifierAction @Inject() (
-  override val authConnector: AuthConnector,
-  pptReferenceAllowedList: PptReferenceAllowedList,
-  override val appConfig: FrontendAppConfig,
-  metrics: Metrics,
-  val parser: BodyParsers.Default
-)(implicit val executionContext: ExecutionContext)
-    extends IdentifierActionOld with AuthorisedFunctions with CommonAuth {
-
-  private val authTimer = metrics.defaultRegistry.timer("ppt.returns.upstream.auth.timer")
-  private val logger    = Logger(this.getClass)
-
-  override def invokeBlock[A](
-    request: Request[A],
-    block: IdentifiedRequest[A] => Future[Result]
-  ): Future[Result] = {
-
-    implicit val hc: HeaderCarrier =
-      HeaderCarrierConverter.fromRequestAndSession(request, request.session)
-
-    // In theory this could be deduplicated with SelectedClientIdentifier by the generic makes it too difficult
-    def getSelectedClientIdentifier() = request.session.get("clientPPT")
-
-    val authorisation            = authTimer.time()
-    val selectedClientIdentifier = getSelectedClientIdentifier()
-
-    authorised(AuthPredicate.createWithEnrolment(selectedClientIdentifier)).retrieve(authData) {
-      case credentials ~ name ~ email ~ externalId ~ internalId ~ affinityGroup ~ allEnrolments ~ agentCode ~
-          confidenceLevel ~ authNino ~ saUtr ~ dateOfBirth ~ agentInformation ~ groupIdentifier ~
-          credentialRole ~ mdtpInformation ~ itmpName ~ itmpDateOfBirth ~ itmpAddress ~ credentialStrength ~ loginTimes =>
-        authorisation.stop()
-
-        val id = internalId.getOrElse(
-          throw new IllegalArgumentException(
-            s"AuthenticatedIdentifierAction::invokeBlock -  internalId is required"
-          )
-        )
-
-        logger.debug(s"AuthenticatedIdentifierAction::invokeBlock - Credentials Provider ID (CredId): ${credentials.map(c => c.providerId)}")
-
-        val identityData = IdentityData(id,
-                                        externalId,
-                                        agentCode,
-                                        credentials,
-                                        Some(confidenceLevel),
-                                        authNino,
-                                        saUtr,
-                                        name,
-                                        dateOfBirth,
-                                        email,
-                                        Some(agentInformation),
-                                        groupIdentifier,
-                                        credentialRole.map(res => res.toJson.toString()),
-                                        mdtpInformation,
-                                        itmpName,
-                                        itmpDateOfBirth,
-                                        itmpAddress,
-                                        affinityGroup,
-                                        credentialStrength,
-                                        Some(loginTimes)
-        )
-
-        affinityGroup match {
-          case Some(AffinityGroup.Agent) =>
-            selectedClientIdentifier.map { pptEnrolmentIdentifier =>
-              // An agent has authed with a selected client and past the enrolment predicate using that identifier
-              // The identifier can be trusted as a good pptEnrolmentIdentifier.
-              executeRequest(request, block, identityData, pptEnrolmentIdentifier, allEnrolments)
-            }.getOrElse {
-              // An agent has authed but we can't see a selected client identifier;
-              // we should prompt them to select one
-              Future.successful(Redirect(agentRoutes.AgentsController.onPageLoad(NormalMode)))
-            }
-
-          case _ =>
-            // A non agent has authed; their ppt enrolment will have been returned
-            // from auth as it is a principal enrolment
-            getPptEnrolmentId(allEnrolments, pptEnrolmentIdentifierName, None) match {
-              case Some(pptEnrolmentIdentifier) =>
-                executeRequest(request, block, identityData, pptEnrolmentIdentifier, allEnrolments)
-              case None =>
-                throw InsufficientEnrolments(
-                  s"key: $pptEnrolmentKey and identifier: $pptEnrolmentIdentifierName is not found"
-                )
-            }
-        }
-    } recover {
-      case _: NoActiveSession =>
-        Redirect(appConfig.loginUrl, Map("continue" -> Seq(request.target.path)))
-
-      case _: InsufficientEnrolments =>
-        // Redirect to the non enrolled page; this is authed but doesn't need enrolments.
-        // There we can examine the user and determine where to send them.
-        Results.Redirect(homeRoutes.UnauthorisedController.notEnrolled())
-
-      case _: AuthorisationException =>
-        Redirect(homeRoutes.UnauthorisedController.unauthorised())
-    }
-  }
-
-  private def executeRequest[A](
-    request: Request[A],
-    block: IdentifiedRequest[A] => Future[Result],
-    identityData: IdentityData,
-    pptEnrolmentIdentifier: String,
-    allEnrolments: Enrolments
-  ) =
-    if (pptReferenceAllowedList.isAllowed(pptEnrolmentIdentifier)) {
-      val pptLoggedInUser = SignedInUser(allEnrolments, identityData)
-      block(IdentifiedRequest(request, pptLoggedInUser, Some(pptEnrolmentIdentifier)))
-    } else {
-      logger.warn("User id is not allowed, access denied")
-      Future.successful(Results.Redirect(homeRoutes.UnauthorisedController.unauthorised()))
-    }
-
-  private def getPptEnrolmentId(
-    enrolments: Enrolments,
-    identifier: String,
-    selectedClientIdentifier: Option[String]
-  ): Option[String] =
-    // It appears Auth with never return a delegated enrolment in it's response to an agent auth request;
-    // therefore we have to use the one the agent past in. This is safe because this identifier has just
-    // past auth for this this agent.
-    maybeClientIdentifier(selectedClientIdentifier).getOrElse(
-      maybePptEnrolmentIdentifier(enrolments, identifier)
-    )
-
-  private def maybeClientIdentifier(
-    selectedClientIdentifier: Option[String]
-  ): Option[Some[String]] =
-    selectedClientIdentifier.map(Some(_))
-
-  private def maybePptEnrolmentIdentifier(
-    enrolments: Enrolments,
-    identifier: String
-  ): Option[String] =
-    getPptEnrolmentIdentifier(enrolments, identifier).filter(_.value.trim.nonEmpty).flatMap(
-      identifier => Some(identifier.value)
-    )
-
-  private def getPptEnrolmentIdentifier(
-    enrolmentsList: Enrolments,
-    identifier: String
-  ): Option[EnrolmentIdentifier] =
-    enrolmentsList.enrolments
-      .filter(_.key == pptEnrolmentKey)
-      .flatMap(_.identifiers)
-      .find(_.key == identifier)
 
 }
