+++ test/base/FakeIdentifierActionWithEnrolment.scala
@@ -17,7 +17,7 @@
 package base
 
 import controllers.actions.IdentifierAction
-import controllers.actions.IdentifierAction.{pptEnrolmentIdentifierName, pptEnrolmentKey}
+import controllers.actions.AuthenticatedIdentifierAction.IdentifierAction.{pptEnrolmentIdentifierName, pptEnrolmentKey}
 import models.SignedInUser
 import models.requests.{IdentifiedRequest, IdentityData}
 import play.api.mvc._
